raw conversion of Disposable specs
